refacto auth
