+++ index.html
@@ -3,9 +3,9 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <script type="module" src='/script/InitialLoadSite.js'></script>
   <script type="module" src='/script/redirect/redirect.js'></script>
   <script type="module" src='/script/validation/authorization.js'></script>
-  <script type="module" src='/script/InitialLoadSite.js'></script>
   <link rel="stylesheet" href="pages/index.css">
   <title>ГосЖалоба</title>
 </head>
