+++ script/index.js
@@ -5,6 +5,7 @@ export {Letter, initialLetters} from './letter/letter.js';
 export {Word, addWord} from './word/word.js';
 export {poems, loadPoems} from './api/loadPoems.js';
 export {getSearchListDefineWords, getDefineWikiSite, getDefine} from './api/wikiApi.js';
+export {enableValidation} from './validation/authorization.js';
 
 export {renderStartPage} from './render/renderStartPage.js';
 export {renderMenu} from './render/renderMenu.js';
