+++ script/redirect/redirect.js
@@ -1,7 +1,2 @@
-import {renderAuthorizationpage, renderStartPage, objHtmlElements} from '../index.js'
 
-if(!localStorage.getItem('test') == '1' || !localStorage.getItem('test') == null) {
-    renderAuthorizationpage(objHtmlElements);
-} else {
-    renderStartPage(objHtmlElements);
-}
+
